Agent class tests finished.
